requirements.txt
alembic==1.8.1
Authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
dlib==19.24.0
dnspython==2.2.1
email-validator==1.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.0
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
